vercelapp_backend/requirements.txt
rav
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-rest-auth
dj-rest-auth[with_social]
Pillow
django-allauth